fix: fix bot response to non-existent users
The bot now only responds once to a non-existent user to inform them that they do not exist on the database. After this, no further correspondence will be issued by the bot on further prompting. If said user is added to the db, they can now prompt the bot freely.